+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Of note:
 - The changelog 2015.5.2 has been rewritten from each commit content.
 - This file may be amended entirely in the future to adhere to the [GNU Changelog style](https://www.gnu.org/prep/standards/html_node/Style-of-Change-Logs.html#Style-of-Change-Logs)
 
+## [2016.12.1]
+### Added
+- Solved [exercice for 2016, day 12](src/year_2016/day_12.rs).
+- First version of [Assembunny](src/year_2016//assembunny.rs) with support for `cpy`, `inc`, `dec`, and `jnz`.
+
 ## [2016.11.1]
 ### Added
 - Solved [exercice for 2016, day 11](src/year_2016/day_11.rs).